refactor(sessions): rename storage.ts to session-store.ts, move interface to contract
