+++ src/cli-chat.ts
@@ -10,8 +10,8 @@ import { t } from "./i18n";
 import { apiUrlForPort, ensureLocalServer } from "./server-daemon";
 import type { Session, SessionState } from "./session-contract";
 import { acquireSessionLock, releaseSessionLock } from "./session-lock";
-import type { SessionStore } from "./session-store";
-import { createSession, getSessionStore } from "./storage";
+import type { SessionStore } from "./session-contract";
+import { createSession, getSessionStore } from "./session-store";
 import { clearScreen, printDim, printError } from "./ui";
 
 const CLI_VERSION = resolveCliVersion();
