+++ src/cli-command-registry.ts
@@ -35,7 +35,7 @@ import {
 } from "./server-daemon";
 import { findSkillByName, loadSkills, readSkillInstructions, SKILL_BUDGET } from "./skills";
 import { formatStatus } from "./status-format";
-import { createSession, getSessionStore } from "./storage";
+import { createSession, getSessionStore } from "./session-store";
 import { openTraceStore } from "./trace-store";
 import { formatCliTitle, printDim, printError, printOutput } from "./ui";
