+++ src/cli-skill.ts
@@ -11,7 +11,7 @@ import { t } from "./i18n";
 import { userResourceIdFor } from "./resource-id";
 import type { apiUrlForPort as apiUrlForPortType, ensureLocalServer as ensureLocalServerType } from "./server-daemon";
 import type { findSkillByName as findSkillByNameType, loadSkills as loadSkillsType } from "./skills";
-import type { createSession as createSessionType } from "./storage";
+import type { createSession as createSessionType } from "./session-store";
 
 type ParsedSkillArgs = { skillName: string; files: string[]; prompt: string; workspace?: string; model?: string };
