+++ src/session-store-contract.test-suite.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, test } from "bun:test";
 import type { Session } from "./session-contract";
-import type { SessionStore } from "./session-store";
+import type { SessionStore } from "./session-contract";
 
 function makeSession(overrides: Partial<Session> & { id: string }): Session {
   return {
