+++ src/session-store-postgres.ts
@@ -2,7 +2,7 @@ import type postgres from "postgres";
 import type { Migration } from "./db-migrate";
 import { log } from "./log";
 import type { Session, SessionId } from "./session-contract";
-import type { SessionStore } from "./session-store";
+import type { SessionStore } from "./session-contract";
 
 type SessionRow = {
   id: string;
