+++ src/session-store.int.test.ts
@@ -1,5 +1,5 @@
 import { sessionStoreContractTests } from "./session-store-contract.test-suite";
-import { createFileSessionStore } from "./storage";
+import { createFileSessionStore } from "./session-store";
 import { tempDb } from "./test-utils";
 
 const { create, cleanup } = tempDb("acolyte-session-", (path) => createFileSessionStore(path));
